requirements.txt
###### Requirements without Version Specifiers ######
python-dotenv 
gql 
requests 
requests_toolbelt 
pydantic 
beautifulsoup4